various missing functions had been added. smal bugs has been solved

+++ bass/core/core.hpp
@@ -56,8 +56,8 @@ struct Plek {
     vector<Frame> frames;
     map<string, CoreFunction> coreFunctions;
     bool strict = false;            // strict mode
-    uint origin = 0;                //file offset
-    int base = 0;                   //file offset to memory map displacement
+    int64_t origin = 0;                //file offset
+    int64_t base = 0;                   //file offset to memory map displacement
     Endian endian = Endian::LSB;    //used for multi-byte writes (d[bwldq], etc)
     Directives directives;          //active directives
     Tracker tracker;                //used to track writes to detect overwrites
